requirements.txt
pandas
pandas_ta
numpy
requests
python-dotenv